requirements.txt
# MCP Server Dependencies
mcp>=1.0.0

# MongoDB
pymongo>=4.6.0

# Async support
aiofiles>=23.2.1

# Environment variables
python-dotenv>=1.0.0

# Date utilities
python-dateutil>=2.8.2
